2つの修正
- useEffectの使用をやめ、処理の複雑化している部分を削除
- queryの初期値の設定
